This version of the OPC UA Server now has an improved node network. Each CIC node now has up to 2 children named after the bus channel to which this CIC is connected. For example: "CIC 0" has the children "CANBus 31" and "CANBus 32". Each bus node itself then has up to 2 children which are the Mops connected to that bus. Each CIC node now also has up to 2 children representing the ADC at it. They are named after the corresponding bus channel. For example: "CIC 0" has the children "CANBus 32" and "CANBus 31" and therefore also "ADC CANBus 32" and "ADC CANBus 31". This node then has children that represent the channel information of the ADC readout. For example: "ADC CANBus32" has the children "Current Monitoring", "GND", "Power Enable", "Temperature", "Voltage Monitoring".
To improve:
The child "Power Enable" will be promoted to a higher level in the tree structure because it is not a value of the ADC.
The find_in_nodes function is not tested at the moment with the new node structure.
The class OPCserver will now generate the MOPSHUBCrate object and then call the SPI/CAN Loop.